--- src/test/java/JUnitFramework/C01_TestNotation.java
@@ -1,34 +1,0 @@
-package JUnitFramework;
-
-import io.github.bonigarcia.wdm.WebDriverManager;
-import org.junit.Test;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
-
-import java.time.Duration;
-
-public class C01_TestNotation {
-
-        @Test
-        public void test01(){
-            WebDriverManager.chromedriver().setup();
-            WebDriver driver = new ChromeDriver();
-
-            driver.manage().window().maximize();
-            driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(15));
-
-            driver.get("https:www.amazon.com");
-
-            WebElement searchBox = driver.findElement(By.id("twotabsearchtextbox"));
-            searchBox.sendKeys("Nutella");
-            searchBox.submit();
-
-            WebElement resultTextElement = driver.findElement(By.xpath("//div[@class='a-section a-spacing-small a-spacing-top-small']"));
-
-            System.out.println(resultTextElement.getText());
-            driver.close();
-        }
-}
-
