+++ src/test/java/seleniumWorks/C02_testNotation.java
@@ -1,4 +1,4 @@
-package JUnitFramework;
+package seleniumWorks;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.junit.Test;
@@ -7,7 +7,7 @@ import org.openqa.selenium.chrome.ChromeDriver;
 
 import java.time.Duration;
 
-public class C02_TestNotation {
+public class C02_testNotation {
 
     @Test
     public void test01(){
